requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
redis==5.0.1
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-exporter-jaeger==1.21.0
sentry-sdk==1.32.0
pydantic==2.4.2
pydantic-settings==2.0.3
tenacity==8.2.3aioredis==2.0.1
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
async-timeout==5.0.1
certifi==2024.12.14
click==8.1.8
Deprecated==1.2.15
exceptiongroup==1.2.2
fastapi==0.104.1
googleapis-common-protos==1.59.1
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.25.1
idna==3.10
importlib-metadata==6.11.0
opentelemetry-api==1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
prometheus-client==0.19.0
protobuf==4.25.5
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.0.1
sentry-sdk==1.32.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tenacity==8.2.3
thrift==0.21.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
wrapt==1.17.0
zipp==3.21.0
